docs: change workflow diagram to list
