+++ www/packages/docs-ui/src/components/WorkflowDiagram/Common/Node/index.tsx
@@ -45,7 +45,7 @@ export const WorkflowDiagramStepNode = ({ step }: WorkflowDiagramNodeProps) => {
       >
         <div
           className={clsx(
-            "shadow-borders-base flex min-w-[120px] bg-medusa-bg-base",
+            "shadow-borders-base flex min-w-[120px] w-min bg-medusa-bg-base",
             "items-center rounded-docs_sm py-docs_0.125 px-docs_0.5",
             (step.type === "hook" || step.when) && "gap-x-docs_0.125"
           )}
